+++ src/hooks/useMapData.js
@@ -115,6 +115,17 @@ export function useMapData(once = false) {
             }
           },
         )
+        const defaultEnabled = prev.filters?.pokemon?.filter?.global?.enabled
+        const serverEnabled = filters.pokemon?.filter?.global?.enabled
+        if (
+          defaultEnabled !== undefined &&
+          serverEnabled !== undefined &&
+          defaultEnabled !== serverEnabled
+        )
+          Object.entries(newFilters.pokemon.filter).forEach(([key, filter]) => {
+            if (prev.filters.pokemon.filter[key] === undefined)
+              filter.enabled = defaultEnabled
+          })
         return {
           filters: newFilters,
         }