+++ src/services/queries/station.js
@@ -11,7 +11,6 @@ const core = gql`
     updated
     start_time
     end_time
-    total_stationed_pokemon
   }
 `
 
@@ -30,6 +29,8 @@ const battle = gql`
     battle_pokemon_bread_mode
     battle_pokemon_move_1
     battle_pokemon_move_2
+    total_stationed_pokemon
+    total_stationed_gmax
   }
 `